Move storage::sets to top level module and remove storage in libsplinter
The only part of the storage module that is still used after the
integration of the AdminServiceStore and RoutingTable is sets::DurableSet.
This commit moves sets to its own top level libsplinter module and removes
the code that is unused in the storage module.
